shop/requirements.txt
boto3==1.26.99
botocore==1.29.99
celery==5.2.7
click==8.1.3
dj-config-url==0.1.1
Django==4.1.7
django-extensions==3.2.1
django-storages==1.13.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Faker==18.3.1
flower==1.2.0
mailhog==0.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
redis==4.5.3
requests==2.28.2
s3transfer==0.6.0
urllib3==1.26.15